==> notebooks/02_process_data.ipynb <==
# Process Web-Scrapped Data

## Imports

import pandas as pd

## Load Data

data_brasil_scola = pd.read_csv('../data/web_scrapped_data/brasil_escola.csv')
kuadro_resume_data = pd.read_csv('../data/web_scrapped_data/kuadro_resume.csv')
me_salva_resumos_data = pd.read_csv('../data/web_scrapped_data/me_salva_resumos.csv')
quero_bolsa_data = pd.read_csv('../data/web_scrapped_data/quero_bolsa.csv')


# create master data, with the columns sub-subject, subject, content, it should be empty at the beginning
master_data = pd.DataFrame(columns=['title', 'sub-subject', 'subject', 'content'])

## Process data_brasil_scola

data_brasil_scola


# remove multiples newlines (including \r and \n) to just one \n
data_brasil_scola['content'] = data_brasil_scola['content'].str.replace(r'[\r\n]+', '\n', regex=True)
# remove PUBLICIDADE
data_brasil_scola['content'] = data_brasil_scola['content'].str.replace('PUBLICIDADE', '')


# convert subject column to lowercase
data_brasil_scola['subject'] = data_brasil_scola['subject'].str.lower()


data_brasil_scola['site'] = 'Brasil Escola'

## Process Kuadro Resume data

# remove multiples \n to just one \n
kuadro_resume_data['content'] = kuadro_resume_data['content'].str.replace('\n+', '\n')

# subject column to lowercase
kuadro_resume_data['subject'] = kuadro_resume_data['subject'].str.lower()

# drop sub-subject column
kuadro_resume_data.drop(columns=['sub-subject'], inplace=True)

# In title, get the entire rest of the string after the first :
kuadro_resume_data['title'] = kuadro_resume_data['title'].str.split(':').str[1]


kuadro_resume_data['site'] = 'Kuadro Resume'

kuadro_resume_data

# ... existing code ...
kuadro_resume_data = kuadro_resume_data[~kuadro_resume_data['title'].str.contains('Gabarito', na=False)]


kuadro_resume_data['subject'].value_counts()

## Process Me Salva Resumos

me_salva_resumos_data['subject'] = me_salva_resumos_data['subject'].str.lower()

me_salva_resumos_data['subject'].value_counts()

me_salva_resumos_data['site'] = 'Me Salva Resumos'

## Process Quero Bolsa Data

# remove title \n
quero_bolsa_data['title'] = quero_bolsa_data['title'].str.replace('\n', '')

# remove \nÍndice\n in the content
quero_bolsa_data['content'] = quero_bolsa_data['content'].str.replace('\nÍndice\n', '')


# remove multiples \n to just one \n
quero_bolsa_data['content'] = quero_bolsa_data['content'].str.replace('\n+', '\n')


quero_bolsa_data['site'] = 'Quero Bolsa'

## Concat Data

master_data = pd.concat([data_brasil_scola, kuadro_resume_data, me_salva_resumos_data, quero_bolsa_data], ignore_index=True)

master_data # do a strip in all columns

master_data['title'] = master_data['title'].str.strip()
master_data['sub-subject'] = master_data['sub-subject'].str.strip()
master_data['subject'] = master_data['subject'].str.strip()
master_data['content'] = master_data['content'].str.strip()

# count nan in all columns
master_data.isna().sum()

# drop nan in content
master_data.dropna(subset=['content'], inplace=True)

# reset index
master_data.reset_index(drop=True, inplace=True)

master_data

# subject column to lowercase
master_data['subject'] = master_data['subject'].str.lower()

# count subject
master_data['subject'].value_counts()



labels_mapping = {'matematica': 'matematica',
          'biologia': 'biologia',
          'fisica': 'fisica',
          'sociologia': 'sociologia',
          'quimica': 'quimica',
          'curiosidades': 'curiosidades',
          'geografia': 'geografia',
          'portugues': 'portugues',
          'filosofia': 'filosofia',
          'literatura': 'literatura',
          'historia': 'historia',
          'ingles': 'ingles',
          'italiano': 'italiano',
          'historia-geral': 'historia',
          'espanhol': 'espanhol',
          'biografias': 'biografias',
          'redacao': 'redacao',
          'historia-brasil': 'historia',
          'artes': 'artes',
          'gramatica': 'gramatica',
          'saude': 'saude',
          'conhecimento-geral': 'conhecimento-geral',
          'historia-da-america': 'historia',
          'biografia': 'biografias',
          'historiab': 'historia',
          'educacao-fisica': 'educacao-fisica',
          'historiag': 'historia',
          'dicas-de-estudo': 'dicas-de-estudo',
          'brasil': 'brasil',
          'acordo-ortografico': 'gramatica'}



# convert subject column using the labels_mapping
master_data['subject'] = master_data['subject'].map(labels_mapping)

master_data['subject'].value_counts()

# drop nan in subject
master_data.dropna(subset=['subject'], inplace=True)

# reset index
master_data.reset_index(drop=True, inplace=True)

# remove data that content is not a str
master_data = master_data[master_data['content'].apply(lambda x: isinstance(x, str))]

# drop nan in content
master_data.dropna(subset=['content'], inplace=True)
# drop empty or strip content
master_data = master_data[master_data['content'].str.strip() != '']

## Save Data

# save to csv
master_data.to_csv('../data/master_data/master_data.csv', index=False)

# load master data
master_data = pd.read_csv('../data/master_data/master_data.csv')

# get master data with nan in content
master_data.content.isna().sum()
