==> notebooks/05_process_gpt_output_data.ipynb <==
# Process output data

## Load Data

import pandas as pd

# load jsonlines 
data_1 = pd.read_json('../gpt_data/output_data/batch_3ilrsK6fo6QrsZz0gCq3knVY_output.jsonl', lines=True)
data_2 = pd.read_json('../gpt_data/output_data/batch_G3rkPDxFiNWXnmk2k5LGmHj2_output.jsonl', lines=True)
data_3 = pd.read_json('../gpt_data/output_data/batch_M1SA4fIDJ4P4GwZ50rykTrYN_output.jsonl', lines=True)
data_4 = pd.read_json('../gpt_data/output_data/batch_NEwqGLKyB6SMatPnTOLVKYFS_output.jsonl', lines=True)


# ... existing code ...
# Concatenate all and then create idx based on custom_id
data = pd.concat([data_1, data_2, data_3, data_4])

# Create idx based on custom_id, handling both formats
data['idx'] = data['custom_id'].apply(lambda x: int(x.split('-')[-1]))

# Create a new column with the request_<idx>
data['request_idx'] = 'request_' + data['idx'].astype(str)

# Set the request_idx as index
data.set_index('request_idx', inplace=True)
# ... existing code ...

data.iloc[0]['response']
# create a new column named gpt_content, from response column 

# create a new column named gpt_content, from response column 
data['gpt_content'] = data['response'].apply(lambda x: x['body']['choices'][0]['message']['content'])

# load master data to do a join with gpt 
master_data = pd.read_csv('../data/master_data/master_data.csv')


master_data

# save as final_data
data.to_csv('../data/final_data/final_data.csv')
